+++ GeometryDashAPI/Levels/GameObjects/Specific/SquareParticle.cs
@@ -7,7 +7,7 @@ namespace GeometryDashAPI.Levels.GameObjects.Specific
     public class SquareParticle : ColorBlock
     {
         [GameProperty("24", (short)Layer.B2)] protected override short zLayer { get; set; } = (short)Layer.B2;
-        [GameProperty("25", (short)0)] public override short ZOrder { get; set; } = 0;
+        [GameProperty("25", 0)] public override int ZOrder { get; set; } = 0;
 
         public SquareParticle() : base(1586)
         {
